fix: corrects alignment issues in sidebar
- also refactors UserLayout to extract NavLink and LanguageSwitcher components
- creates an IconContainer to maintain menu icon dimensions
