+++ src/features/common/Layout/UserLayout/LanguageSwitcher.tsx
@@ -1,0 +1,67 @@
+import { useState, useEffect } from 'react';
+import { useLocale } from 'next-intl';
+import SelectLanguageAndCountry from '../Footer/SelectLanguageAndCountry';
+import GlobeIcon from '../../../../../public/assets/images/icons/Sidebar/Globe';
+import IconContainer from './IconContainer';
+import styles from './UserLayout.module.scss';
+
+function LanguageSwitcher() {
+  const locale = useLocale();
+
+  const [language, setLanguage] = useState(locale);
+  const [openModal, setOpenModal] = useState(false);
+  const [selectedCurrency, setSelectedCurrency] = useState('EUR');
+  const [selectedCountry, setSelectedCountry] = useState('DE');
+
+  useEffect(() => {
+    if (typeof Storage !== 'undefined') {
+      //fetching language from browser's local storage
+      if (localStorage.getItem('language')) {
+        const langCode = localStorage.getItem('language') || 'en';
+        if (langCode) setLanguage(langCode.toLowerCase());
+      }
+    }
+  }, [language]);
+
+  useEffect(() => {
+    if (typeof Storage !== 'undefined') {
+      //fetching currencyCode from browser's localstorage
+      if (localStorage.getItem('currencyCode')) {
+        const currencyCode = localStorage.getItem('currencyCode');
+        if (currencyCode) setSelectedCurrency(currencyCode);
+      }
+      //fetching country code from browser's localstorage
+      if (localStorage.getItem('countryCode')) {
+        const countryCode = localStorage.getItem('countryCode');
+        if (countryCode) setSelectedCountry(countryCode);
+      }
+    }
+  }, []);
+
+  return (
+    <>
+      <div className={styles.navLink}>
+        <IconContainer>
+          <GlobeIcon />
+        </IconContainer>
+        <button
+          className={styles.navLinkTitle}
+          onClick={() => {
+            setOpenModal(true); // open language and country change modal
+          }}
+        >
+          {`${locale ? locale.toUpperCase() : ''} • ${selectedCurrency}`}
+        </button>
+      </div>
+      <SelectLanguageAndCountry
+        openModal={openModal}
+        handleModalClose={() => setOpenModal(false)}
+        setSelectedCurrency={setSelectedCurrency}
+        selectedCountry={selectedCountry}
+        setSelectedCountry={setSelectedCountry}
+      />
+    </>
+  );
+}
+
+export default LanguageSwitcher;
