+++ src/features/common/Layout/UserLayout/UserLayout.tsx
@@ -1,14 +1,12 @@
-import type { User } from '@planet-sdk/common/build/types/user';
-import type { Dispatch, FC, ReactNode, SetStateAction } from 'react';
+import type { FC } from 'react';
+import type { NavLinkType, SubMenuItemType } from './NavLink';
 
-import router, { useRouter } from 'next/router';
+import { useRouter } from 'next/router';
 import React, { useEffect, useState } from 'react';
 import { useLocale, useTranslations } from 'next-intl';
 import MenuIcon from '../../../../../public/assets/images/icons/Sidebar/MenuIcon';
-import DownArrow from '../../../../../public/assets/images/icons/DownArrow';
 import BackArrow from '../../../../../public/assets/images/icons/headerIcons/BackArrow';
 import DonateIcon from '../../../../../public/assets/images/icons/Sidebar/DonateIcon';
-import GlobeIcon from '../../../../../public/assets/images/icons/Sidebar/Globe';
 import LogoutIcon from '../../../../../public/assets/images/icons/Sidebar/LogoutIcon';
 import MapIcon from '../../../../../public/assets/images/icons/Sidebar/MapIcon';
 import PlanetCashIcon from '../../../../../public/assets/images/icons/Sidebar/PlanetCashIcon';
@@ -16,7 +14,6 @@ import SettingsIcon from '../../../../../public/assets/images/icons/Sidebar/Sett
 import UserIcon from '../../../../../public/assets/images/icons/Sidebar/UserIcon';
 import WidgetIcon from '../../../../../public/assets/images/icons/Sidebar/Widget';
 import { UserProfileLoader } from '../../ContentLoaders/UserProfile/UserProfile';
-import SelectLanguageAndCountry from '../Footer/SelectLanguageAndCountry';
 import { useUserProps } from '../UserPropsContext';
 import styles from './UserLayout.module.scss';
 import TreeMapperIcon from '../../../../../public/assets/images/icons/Sidebar/TreeMapperIcon';
@@ -24,195 +21,9 @@ import RegisterTreeIcon from '../../../../../public/assets/images/icons/Sidebar/
 import NotionLinkIcon from '../../../../../public/assets/images/icons/Sidebar/NotionLinkIcon';
 import SupportPin from '../../../user/Settings/ImpersonateUser/SupportPin';
 import FiberPinIcon from '@mui/icons-material/FiberPin';
-
-interface SubMenuItemType {
-  title: string;
-  path: string;
-  flag?: string;
-  hideItem?: boolean;
-}
-
-interface NavLinkType {
-  key: number;
-  title: string;
-  path?: string;
-  icon: ReactNode;
-  flag?: string;
-  accessLevel?: string[];
-  hideSubMenu?: boolean;
-  subMenu?: SubMenuItemType[];
-  hideItem?: boolean;
-  hasRelatedLinks?: boolean;
-}
-
-function LanguageSwitcher() {
-  const locale = useLocale();
-
-  const [language, setLanguage] = useState(locale);
-  const [openModal, setOpenModal] = useState(false);
-  const [selectedCurrency, setSelectedCurrency] = useState('EUR');
-  const [selectedCountry, setSelectedCountry] = useState('DE');
-
-  useEffect(() => {
-    if (typeof Storage !== 'undefined') {
-      //fetching language from browser's local storage
-      if (localStorage.getItem('language')) {
-        const langCode = localStorage.getItem('language') || 'en';
-        if (langCode) setLanguage(langCode.toLowerCase());
-      }
-    }
-  }, [language]);
-
-  useEffect(() => {
-    if (typeof Storage !== 'undefined') {
-      //fetching currencyCode from browser's localstorage
-      if (localStorage.getItem('currencyCode')) {
-        const currencyCode = localStorage.getItem('currencyCode');
-        if (currencyCode) setSelectedCurrency(currencyCode);
-      }
-      //fetching country code from browser's localstorage
-      if (localStorage.getItem('countryCode')) {
-        const countryCode = localStorage.getItem('countryCode');
-        if (countryCode) setSelectedCountry(countryCode);
-      }
-    }
-  }, []);
-
-  return (
-    <>
-      <div className={styles.navLink}>
-        <GlobeIcon />
-        <button
-          className={styles.navLinkTitle}
-          onClick={() => {
-            setOpenModal(true); // open language and country change modal
-          }}
-        >
-          {`${locale ? locale.toUpperCase() : ''} • ${selectedCurrency}`}
-        </button>
-      </div>
-      <SelectLanguageAndCountry
-        openModal={openModal}
-        handleModalClose={() => setOpenModal(false)}
-        setSelectedCurrency={setSelectedCurrency}
-        selectedCountry={selectedCountry}
-        setSelectedCountry={setSelectedCountry}
-      />
-    </>
-  );
-}
-interface NavLinkProps {
-  link: NavLinkType;
-  setActiveLink: Dispatch<SetStateAction<string>>;
-  activeLink: string;
-  activeSubMenu: string;
-  setActiveSubMenu: Dispatch<SetStateAction<string>>;
-  user: User;
-  key: number;
-  closeMenu: () => void;
-}
-function NavLink({
-  link,
-  setActiveLink,
-  activeLink,
-  activeSubMenu,
-  setActiveSubMenu,
-  user,
-}: NavLinkProps) {
-  const [isSubMenuActive, setIsSubMenuActive] = useState(false);
-  const locale = useLocale();
-  useEffect(() => {
-    // Check if array of submenu has activeSubLink
-    if (link.subMenu && link.subMenu.length > 0) {
-      const subMenuItem = link.subMenu.find((subMenuItem) => {
-        return subMenuItem.path === activeSubMenu;
-      });
-      if (subMenuItem) {
-        link.path && setActiveLink(link.path);
-        setActiveSubMenu(subMenuItem.path);
-        if (activeSubMenu && subMenuItem.path === activeSubMenu) {
-          setIsSubMenuActive(true);
-        }
-      }
-    }
-  }, [activeLink]);
-
-  if (link.accessLevel) {
-    //checks the type of user login
-    if (!link.accessLevel.includes(user.type)) {
-      return null;
-    }
-  }
-
-  return (
-    <div key={link.title} className={styles.navLinkMenu}>
-      <div
-        className={`${styles.navLink} ${
-          activeLink && activeLink === link.path ? styles.navLinkActive : ''
-        } ${isSubMenuActive ? styles.navLinkActive : ''}`}
-        onClick={() => {
-          // This is to shift to the main page needed when there is no sub menu
-          if ((!link.subMenu || link.subMenu.length <= 0) && link.path) {
-            router.push(`/${locale}${link.path}`);
-            setActiveLink(link.path);
-            setActiveSubMenu('');
-          } else {
-            if (link.hideSubMenu && link.path) {
-              router.push(`/${locale}${link.path}`);
-            } else {
-              setIsSubMenuActive(!isSubMenuActive);
-            }
-          }
-        }}
-      >
-        {link.icon}
-        <button className={styles.navLinkTitle}>
-          {link.title}
-          {link.flag && <span className={styles.itemFlag}>{link.flag}</span>}
-        </button>
-        {link.subMenu && link.subMenu.length > 0 && !link.hideSubMenu && (
-          <button
-            className={styles.subMenuArrow}
-            style={{
-              transform: isSubMenuActive ? 'rotate(-180deg)' : 'rotate(-90deg)',
-            }}
-          >
-            <DownArrow />
-          </button>
-        )}
-      </div>
-      {isSubMenuActive &&
-        link.subMenu &&
-        link.subMenu.length > 0 &&
-        !link.hideSubMenu &&
-        link.subMenu.map((subLink, index) => {
-          if (!subLink.hideItem) {
-            return (
-              <div
-                className={`${styles.navLinkSubMenu} ${
-                  activeSubMenu === subLink.path
-                    ? styles.navLinkActiveSubMenu
-                    : ''
-                }`}
-                key={index}
-                onClick={() => {
-                  //this is to shift to the submenu pages
-                  link.path && setActiveLink(link.path);
-                  setActiveSubMenu(subLink.path);
-                  router.push(`/${locale}${subLink.path}`);
-                }}
-              >
-                {subLink.title}
-                {subLink.flag && (
-                  <span className={styles.itemFlag}>{subLink.flag}</span>
-                )}
-              </div>
-            );
-          }
-        })}
-    </div>
-  );
-}
+import IconContainer from './IconContainer';
+import LanguageSwitcher from './LanguageSwitcher';
+import NavLink from './NavLink';
 
 const UserLayout: FC = ({ children }) => {
   const t = useTranslations('Me');
@@ -222,7 +33,6 @@ const UserLayout: FC = ({ children }) => {
     useUserProps();
 
   // Flags can be added to show labels on the right
-  // TO DO - remove arrow when link is selected
   const navLinks: NavLinkType[] = [
     {
       key: 1,
@@ -452,12 +262,16 @@ const UserLayout: FC = ({ children }) => {
 
           {!isImpersonationModeOn && (
             <div className={styles.navLink}>
-              <FiberPinIcon />
+              <IconContainer>
+                <FiberPinIcon />
+              </IconContainer>
               <SupportPin />
             </div>
           )}
           <div className={styles.navLink}>
-            <NotionLinkIcon />
+            <IconContainer>
+              <NotionLinkIcon />
+            </IconContainer>
             <button
               onClick={() =>
                 window.open(
@@ -478,9 +292,10 @@ const UserLayout: FC = ({ children }) => {
               logoutUser(`${window.location.origin}/`);
             }}
           >
-            <LogoutIcon />
+            <IconContainer>
+              <LogoutIcon />
+            </IconContainer>
             <button className={styles.navLinkTitle}>{t('logout')}</button>
-            <button className={styles.subMenuArrow}></button>
           </div>
         </div>
       </div>
